refactor(format): extract analysis xml renderer
